requirements.txt
certifi==2019.9.11
olefile==0.46
Pillow==6.2.0
wincertstore==0.2
tk==8.6.8
